changed Assembler

+++ sim/fpu.cpp
@@ -3,7 +3,7 @@ namespace StdFPU{
 }
 
 namespace OrenoFPU{
-    long long sqrt_grad[1024] = {
+    int64_t sqrt_grad[1024] = {
     0x6a09f1699, 0x6a6467693, 0x6abec968e, 0x6b1913688, 
     0x6b7347683, 0x6bcd6367d, 0x6c276b677, 0x6c815b672, 
     0x6cdb3566c, 0x6d34fb667, 0x6d8ea9661, 0x6de84165c, 
@@ -262,7 +262,7 @@ namespace OrenoFPU{
     0x6954b8b53, 0x69820cb51, 0x69af5cb50, 0x69dca6b4e
     };
 
-    long long div_grad[1024] = {
+    int64_t div_grad[1024] = {
     0xfffffbff7, 0xff801bfe8, 0xff007bfd8, 0xfe8119fc8, 
     0xfe01f9fb8, 0xfd8317fa8, 0xfd0475f98, 0xfc8611f89, 
     0xfc07ebf79, 0xfb8a05f6a, 0xfb0c5bf5a, 0xfa8ef1f4b,
